moves all unit tests to "unit" package and create Suites
