+++ core/src/test/java/org/aerogear/mobile/core/integration/MetricsServiceIntegrationTest.java
@@ -1,0 +1,22 @@
+package org.aerogear.mobile.core.integration;
+
+import static org.junit.Assert.fail;
+
+import org.junit.Before;
+import org.junit.Test;
+import org.junit.runner.RunWith;
+
+@RunWith(PowerMockRunner.class)
+public class MetricsServiceIntegrationTest {
+
+    @Before
+    public void setUp() throws Exception {
+
+    }
+
+    @Test
+    public void testSendMetricsShouldReturnNoError() throws Exception {
+        fail("Not yet implemented");
+    }
+
+}
