+++ core/src/test/java/org/aerogear/mobile/core/unit/configuration/ServiceConfigurationTest.java
@@ -1,9 +1,11 @@
-package org.aerogear.mobile.core.configuration;
+package org.aerogear.mobile.core.unit.configuration;
 
 import org.junit.Assert;
 import org.junit.Test;
 import org.junit.runner.RunWith;
 import org.robolectric.RobolectricTestRunner;
+
+import org.aerogear.mobile.core.configuration.ServiceConfiguration;
 
 @RunWith(RobolectricTestRunner.class)
 public class ServiceConfigurationTest {
