+++ core/src/test/java/org/aerogear/mobile/core/unit/metrics/impl/DeviceMetricsTest.java
@@ -1,4 +1,4 @@
-package org.aerogear.mobile.core.metrics.impl;
+package org.aerogear.mobile.core.unit.metrics.impl;
 
 import static org.junit.Assert.assertEquals;
 import static org.junit.Assert.assertNotNull;
@@ -12,6 +12,8 @@ import org.robolectric.RuntimeEnvironment;
 
 import android.app.Application;
 import android.support.test.filters.SmallTest;
+
+import org.aerogear.mobile.core.metrics.impl.DeviceMetrics;
 
 @RunWith(RobolectricTestRunner.class)
 @SmallTest
