+++ core/src/test/java/org/aerogear/mobile/core/unit/utils/SanityCheckTest.java
@@ -1,4 +1,4 @@
-package org.aerogear.mobile.core.utils;
+package org.aerogear.mobile.core.unit.utils;
 
 import org.junit.Assert;
 import org.junit.Test;
@@ -6,6 +6,8 @@ import org.junit.runner.RunWith;
 import org.robolectric.RobolectricTestRunner;
 
 import android.support.test.filters.SmallTest;
+
+import org.aerogear.mobile.core.utils.SanityCheck;
 
 @RunWith(RobolectricTestRunner.class)
 @SmallTest
